requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
openai-whisper>=20231117
Werkzeug>=2.3.0
ffmpeg-python>=0.2.0
python.dotenv>=0.19.0
openai>=1.0.0
